+++ rules.tf
@@ -111,6 +111,29 @@ resource "azurerm_network_security_group" "bastion" {
     destination_address_prefix  = "Internet"
   }
 
+  dynamic "security_rule" {
+    for_each = var.security_rules
+    content {
+      name        = security_rule.key
+      priority    = security_rule.value.priority
+      direction   = security_rule.value.direction
+      access      = security_rule.value.access
+      protocol    = security_rule.value.protocol
+
+      source_port_range  = try(security_rule.value.source_port_range, null)
+      source_port_ranges = try(security_rule.value.source_port_ranges, null)
+
+      destination_port_range  = try(security_rule.value.destination_port_range, null)
+      destination_port_ranges = try(security_rule.value.destination_port_ranges, null)
+
+      source_address_prefix  = try(security_rule.value.source_address_prefix, null)
+      source_address_prefixes = try(security_rule.value.source_address_prefixes, null)
+
+      destination_address_prefix  = try(security_rule.value.destination_address_prefix, null)
+      destination_address_prefixes = try(security_rule.value.destination_address_prefixes, null)
+    }
+  }
+
   security_rule {
     name                       = var.bastion_subnet_id != null ? "Deny-${local.subnet_numeric}-to-${local.subnet_numeric}-Any-Inbound" : "Deny-${local.new_subnet_numeric}-to-${local.new_subnet_numeric}-Any-Inbound"
     priority                   = 4096